Drop the I/O timeout down to three hours.
At the moment we kill the rsync if there is no I/O for the given
period (6 hours previously, 3 after this commit lands) and then
retry the rsync connection up to three times.

This meant that a backup could take no more than 3x6 = 18 hours
but that was too long.  Now we'll be 3x2 = 6 hours and that should
be sufficient that transient failures are handled, but
never-gonna-complete cases fail in a timely fashion.